requirements.txt
djangogrpcframework
django==3.1
djangorestframework==3.11.0
grpcio
grpcio-tools
psycopg2-binary
protobuf